[2127] Addressed PR comments for Root activity and its components
